Issue #1405: Rename Tool util class to Docker
